move text-based model to helper function
